Clean up clang warnings. (#971)

+++ src/selfplay/game.cc
@@ -26,6 +26,7 @@
 */
 
 #include "selfplay/game.h"
+
 #include <algorithm>
 
 #include "neural/writer.h"
@@ -180,8 +181,9 @@ void SelfPlayGame::Play(int white_threads, int black_threads, bool training,
       }
       // If 'best move' is less than allowed visits and not max visits,
       // discard it and try again.
-      if (cur_n == max_n || cur_n >= options_[idx].uci_options->Get<int>(
-                                        kMinimumAllowedVistsId.GetId())) {
+      if (cur_n == max_n ||
+          static_cast<int>(cur_n) >= options_[idx].uci_options->Get<int>(
+                                         kMinimumAllowedVistsId.GetId())) {
         break;
       }
       auto move_list_to_discard = GetMoves();
